Debugged errors caused by session id validation
IP address and session id validation was added in the last commit (b2ed1e4) to ensure that socket connection could only take place from the same IP address and session from which POST/ registerKeys request was made. However, this caused a problem, because the session was not being shared between express and socket.io, causing the server to reject the connection, which caused the client to refresh endlessly.

 - The problem with the same session not being shared between express and socket.io was resolved by setting 'resave': true and 'saveUninitialized': true. This is a temporary solution: will later stop using express-session and instead rely on custom cookie to track and enforce sessions.

 - Server now sends fresh js file hash on socket handshake response so that the client refreshes *only* if its hash doesn't match the provided one, preventing endless refresh.