Removed the 'buffer' dependency
The npm module 'buffer' was being used as a stand-in for 'node:buffer' throughout the app, on both the backend and the frontend. This required conversion to and from 'node:buffer' where it was explicitly required (mainly mongoose).

All uses of npm 'buffer' in the code have now been removed, using 'node:buffer' everywhere, and it is no longer a dependency anywhere. It is just provided as a webpack polyfill to frontend code (as a dev-dependency).