Push notifications for ChatRequests added
Before the restructuring of the database schema in the previous commit ba03339,  clients were notified of any ChatRequests, or accepted ChatRequests proactively by the server as they arrived. However, since in ba03339, most stored data was no longer connected to username, the push notification implementation became defunct and was removed.

In this message, a new implementation for push notification is introduced, which is compatible with the new schema and requires almost no change in the frontend. Online clients are now notified when ChatRequests addressed to them are sent, or when ChatRequests sent by them are accepted. This also includes a fallback mechanism that notifies clients if MessageHeaders or Receipts arrive through the database rather than a socket channel (Room).