EncryptionVector derivation and schema changed
Previously, some core data belonging to the client was stored in a password-encrypted form, i.e., encrypted by a password-derived scalar and inaccessible by the server. The data included encryptionBaseVector (used to encrypt all the user data, chats, messages, etc.), databaseAuthKey (used to verify access to database records), clientIdentityVerifyingKey (used to sign data sent by client to the server), and serverIdentityVerifyingKey (used to verify data sent by the server). Each of these was encrypted by a different password-derived scalar, together with the parameters used to derive it.

This process has been completely restructured. Now, no information is stored in a password-encrypted form. Instead, encryptionBaseVector and databaseAuthKey are expressed as randomly generated public points on ed25519 (using Ristretto-mapped points). They are then entangled with separate password-derived scalars (each derived with a different set of parameters). The password-entangled public points for encryptionBaseVector and databaseAuthKey together with the parameters used to derive the scalars are sent to and stored by the server. Everything else now (including clientIdentitySigningKey and serverIdentityVerifyingKey) is encrypted using the encryptionBaseVector, which is not available to the database.

During the logging in process, the client is provided first with the stored parameters and password-entangled point for databaseAuthKey. The client then uses the parameters to derive the original scalar, computes its modular inverse over the order of ed25519, and multiplies the entangled point by it, disentangling it and recovering the original random point, which is used to derive databaseAuthKeyBuffer, which is sent to the server to verify access to appropriate database records (through DatabaseAccessKeys). If the client is able to successfully authenticate itself, it is also provided with the stored parameters and entangled point for encryptionBaseVector, for which it repeats the above process, recovering encryptionBaseVector, which is then used to decrypt the rest of the data provided by the server.

Moreover, the scalar (verifierPrivate) which is used to calculate verifierPoint and authenticate user is now calculated similarly as the rest of the password-derived scalars.

Also, since clientIdentitySigningKey and serverIdentityVerifyingKey are no longer password-encrypted, they are no longer stored in local storage together with encryptionBaseVector when saving password, and are instead provided by the server on login. However, databaseAuthKeyBuffer *is* stored in local storage, since it also depends on the password.

Minor changes have been made to ellipticSRP (other than the restructuring), which involve the full use of Buffers as input and output instead of hex strings.

Minor changes made to cryptoOperator to streamline encryption/decryption of private keys.