Changed how js file hashes work
Previously, the server sent the latest hash of main.js to the client in response to every request. The client then compared the file hash with its calculated hash of the running file, and reloaded if it they didn't match. However, this was inefficient.

Now, the server writes the latest main.js hash to a publicly available txt file, updating it whenever main.js changes. The client uses setInterval to read the latest hash from the txt file every second and compare it with its hash of the running file. If the hashes don't match, the client reloads. This way is better with respect to separation of concerns, and allows the client to reload with less delay when the file changes.