Authentication reimplemented with REST on backend
Previously, user login and signup operations took place over the socket, and were handled by the SocketHandler, which was initially created in an unauthenticated state.

Now, the methods responsible for login and signup have all been shifted to a separate class (AuthHandler). The communication for logging in or signing up is no longer carried over the socket connection, and the socket connection cannot be formed before the client has logged in, which is now done through two REST methods, GET and CONNECT. When the user has logged in, the authenticated cookie is set with the sessionReference, and a SocketHandler is created. When the client subsequently attempts to create a socket connection, the authenticated cookie is checked to see if the connection should be allowed, and then attaches the connected socket to the SocketHandler created at the conclusion of the authentication process (through the sessionReference).

Minor changes:

- SessionCrypto now concatenates its encryption purpose string with its sessionReference to make the encrypted data strongly tied to the session.