RunningClientSession storage is now encrypted
In the previous commit (dc0b35f), the ability to persist session information was implemented, so that if the server restarted, it could read the data for any sessions running before the crash and resume those sessions without the client having to re-authenticate itself.

However, this session information was being stored in the database in transparent way (containing username, sharedSecret, etc.), so even though the session record was temporary and only lasted as long as the session, sensitive information could be determined from the database state during that time.

Now, when logging in or signing up, a sessionRecordKey is derived from SharedKeyBits using a random salt on the backend. The salt is sent to the client if login/signup succeeds. The sessionRecordKey is used to encrypt the session info, which is only stored to the database in its encrypted form, identified only by sessionReference. The sessionRecordKey is not retained by the backend.

The client, when receiving the salt used to derive sessionRecordKey upon login/signup, uses it to derive the sessionRecordKey itself, and keeps it. Then, when the client attempts to confirm its authentication or establish a socket connection, it must also supply the sessionRecordKey, which the server will use to decrypt the session record from the database (if needed) before confirming authentication/establishing connection.

This way, the session information stored in the database is inaccessible without the client sending the key to decrypt it.