Logging in from same device multiple times handled
Previously, while logged in on one device, the user would be barred from logging in on another device. However, if the user attempted to log in in a new tab on the same device, the old session would be disrupted.

Now, when logging in from a new tab on the same device, the user is logged out of the session from the old tab without showing "Connection down".