Authenticated session now persists across reloads
When a client logs in, it is provided with a random bit sequence, which it uses to encrypt information about the authenticated session and store it in SessionStorage of the browser.

When the client sends a POST/terminateCurrentSession request, instead of disposing the session right away, the server sets up a timeout to dispose it in 5 seconds.

On loading, a client will check SessionStorage. If a SavedSession is found, it will send a GET/resumeAuthenticatedSession. If the server finds a timeout, it will clear the timeout and send back the saveSessionKey, which the client can use to decrypt the information from SessionStorage and recreate the authenticated session.