Fixed bug in useScrollRestore
Fixed a bug that probably arose due to changes made to the onOrientationChange handler, which was causing the message at top to change when rotating. The cause was the scroll position being calculated at a time different from when it was being set. Moving the calculation (back) inside the setTimeout solved the issue.

Also fixed a potential bug where a rotation might have caused more messages to load, by checking if the orientation has changed before loading more messages.